feat (voucher): logic update bill

+++ src/app/api/voucher/[id]/route.ts
@@ -5,25 +5,39 @@ import { Voucher } from '@/data/types';
 
 const filePath = path.resolve(process.cwd(), 'src/data/vouchers.json');
 
-// [PUT] with api/voucher/[id]
 export async function PUT(req: NextRequest, context: { params: { id: string } }) {
-  const { id } = await context.params;
+  const { id } = context.params;
   const updatedVoucher: Voucher = await req.json();
 
   let vouchers = JSON.parse(fs.readFileSync(filePath, 'utf8')) as Voucher[];
-  vouchers = vouchers.map(v => v.id === id ? updatedVoucher : v);
+  const index = vouchers.findIndex(v => v.id === id);
 
+  if (index === -1) {
+    return NextResponse.json(
+      { message: 'Voucher not found' },
+      { status: 404 }
+    );
+  }
+
+  vouchers[index] = updatedVoucher;
   fs.writeFileSync(filePath, JSON.stringify(vouchers, null, 2));
   return NextResponse.json({ message: 'Voucher updated' });
 }
 
-// [DELETE] with api/voucher/[id]
 export async function DELETE(_: NextRequest, context: { params: { id: string } }) {
-  const { id } = await context.params;
+  const { id } = context.params;
 
   let vouchers = JSON.parse(fs.readFileSync(filePath, 'utf8')) as Voucher[];
-  vouchers = vouchers.filter(v => v.id !== id);
+  const exists = vouchers.find(v => v.id === id);
 
+  if (!exists) {
+    return NextResponse.json(
+      { message: 'Voucher not found' },
+      { status: 404 }
+    );
+  }
+
+  vouchers = vouchers.filter(v => v.id !== id);
   fs.writeFileSync(filePath, JSON.stringify(vouchers, null, 2));
   return NextResponse.json({ message: 'Voucher deleted' });
 }
